requirements.txt
Flask
pytesseract
Pillow 